fix: refactor feature values and types
