+++ packages/shared/src/components/cards/PostCard.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, RenderResult, screen, waitFor } from '@testing-library/react';
 import { Post } from '../../graphql/posts';
 import { PostCard, PostCardProps } from './PostCard';
-import { AdditionalInteractionButtons } from '../../lib/featureManagement';
+import { AdditionalInteractionButtons } from '../../lib/featureValues';
 
 const defaultPost: Post = {
   id: 'e3fd75b62cadd02073a31ee3444975cc',
