+++ packages/shared/src/lib/featureValues.ts
@@ -1,0 +1,4 @@
+export enum AdditionalInteractionButtons {
+  Bookmark = 'bookmark',
+  Share = 'share',
+}
